+++ Navigation/ProfileNavigation.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import ProfilePage from '../screens/ProfilePage';
-import UpdateProfilePage from '../screens/UpdateProfilePage'
+import ProfilePage from '../screens/profile/ProfilePage';
+import UpdateProfilePage from '../screens/profile/UpdateProfilePage'
 
 const Stack = createNativeStackNavigator();
