+++ screens/chat/ParticipantsPage.js
@@ -8,7 +8,7 @@ import {
     FlatList,
     Image,
 } from 'react-native';
-import { supabase } from '../supabase';
+import { supabase } from '../../supabase';
 import { useRoute } from '@react-navigation/native';
 import { set } from 'react-native-reanimated';
