+++ screens/match/SpecificMatchPage.js
@@ -7,7 +7,7 @@ import {
     SafeAreaView,
     FlatList,
 } from 'react-native';
-import { supabase } from '../supabase';
+import { supabase } from '../../supabase';
 
 const SpecificMatchPage = () => {
