+++ screens/request/ViewOwnRequestPage.js
@@ -7,8 +7,8 @@ import {
     Text,
     FlatList,
   } from 'react-native';
-import { supabase } from '../supabase';
-import Request from '../components/Request';
+import { supabase } from '../../supabase';
+import Request from '../../components/request/Request';
 
 const ViewOwnRequestPage = () => {
 
@@ -37,20 +37,8 @@ const ViewOwnRequestPage = () => {
   const getRequests = async () => {
     try {
       const { data, error } = await supabase.from('Requests')
-      .select(`
-      id,
-      requestor_id,
-      username:profiles (Username),
-      created_at,
-      Location,
-      Time,
-      Date,
-      Pax,
-      Description,
-      Title,
-      Request_url
-      `).eq('requestor_id', user.id)
-      .order('Date', { ascending: true }).order('Time', { ascending: true });
+      .select(`*, username:profiles (Username)`).eq('requestor_id', user.id)
+      .order('datetime', { ascending: true });
       if (error) throw error
       setYourRequests(data)
     } catch (error) {
